refactor(sdk/go): split models per-resource, extract shared types
Move meter models into meters.go alongside MetersService; extract
shared types into types.go (Numeric), pagination.go (PageParams,
PageMeta, PaginatedMeta), and filters.go (StringFilter). Remove
models.go. No behavior change; new resources now add one file each.
