+++ sdk/go/openmeter/meters.go
@@ -8,6 +8,7 @@ import (
 	"net/url"
 	"strconv"
 	"strings"
+	"time"
 )
 
 const metersBasePath = "/openmeter/meters"
@@ -16,41 +17,137 @@ const metersBasePath = "/openmeter/meters"
 // specify one, chosen to keep round-trips low without oversized responses.
 const defaultListPageSize = 100
 
+// MeterAggregation is the aggregation type a meter applies to matched events.
+type MeterAggregation string
+
+const (
+	MeterAggregationSum         MeterAggregation = "sum"
+	MeterAggregationCount       MeterAggregation = "count"
+	MeterAggregationUniqueCount MeterAggregation = "unique_count"
+	MeterAggregationAvg         MeterAggregation = "avg"
+	MeterAggregationMin         MeterAggregation = "min"
+	MeterAggregationMax         MeterAggregation = "max"
+	MeterAggregationLatest      MeterAggregation = "latest"
+)
+
+// Meter is a configuration that defines how to match and aggregate events.
+type Meter struct {
+	ID          string           `json:"id"`
+	Key         string           `json:"key"`
+	Name        string           `json:"name"`
+	Description *string          `json:"description,omitempty"`
+	Aggregation MeterAggregation `json:"aggregation"`
+	EventType   string           `json:"event_type"`
+	// EventsFrom, when set, is the date from which the meter includes events.
+	EventsFrom *time.Time `json:"events_from,omitempty"`
+	// ValueProperty is a JSONPath expression extracting the aggregated value from
+	// the event data. Ignored for count aggregation.
+	ValueProperty *string `json:"value_property,omitempty"`
+	// Dimensions maps group-by dimension names to JSONPath expressions.
+	Dimensions map[string]string `json:"dimensions,omitempty"`
+	Labels     map[string]string `json:"labels,omitempty"`
+	CreatedAt  time.Time         `json:"created_at"`
+	UpdatedAt  time.Time         `json:"updated_at"`
+	DeletedAt  *time.Time        `json:"deleted_at,omitempty"`
+}
+
+// MeterPagePaginatedResponse is a page of meters plus pagination metadata.
+type MeterPagePaginatedResponse struct {
+	Data []Meter       `json:"data"`
+	Meta PaginatedMeta `json:"meta"`
+}
+
+// MeterQueryGranularity is the size of the time buckets a query groups usage into.
+type MeterQueryGranularity string
+
+const (
+	MeterQueryGranularityMinute MeterQueryGranularity = "PT1M"
+	MeterQueryGranularityHour   MeterQueryGranularity = "PT1H"
+	MeterQueryGranularityDay    MeterQueryGranularity = "P1D"
+	MeterQueryGranularityMonth  MeterQueryGranularity = "P1M"
+)
+
+// QueryFilterStringMapItem is a per-dimension filter in a meter query. For the
+// reserved subject and customer_id dimensions only Eq/In are supported.
+type QueryFilterStringMapItem struct {
+	Eq        *string  `json:"eq,omitempty"`
+	Neq       *string  `json:"neq,omitempty"`
+	In        []string `json:"in,omitempty"`
+	Nin       []string `json:"nin,omitempty"`
+	Contains  *string  `json:"contains,omitempty"`
+	Ncontains *string  `json:"ncontains,omitempty"`
+	Exists    *bool    `json:"exists,omitempty"`
+}
+
+// MeterQueryFilters filters a meter query by dimension values.
+type MeterQueryFilters struct {
+	Dimensions map[string]QueryFilterStringMapItem `json:"dimensions,omitempty"`
+}
+
+// MeterQueryRequest is the POST body for querying a meter for usage.
+type MeterQueryRequest struct {
+	From        *time.Time             `json:"from,omitempty"`
+	To          *time.Time             `json:"to,omitempty"`
+	Granularity *MeterQueryGranularity `json:"granularity,omitempty"`
+	// TimeZone is an IANA Time Zone Database name used to align time buckets.
+	// Defaults to UTC when unset.
+	TimeZone          *string            `json:"time_zone,omitempty"`
+	GroupByDimensions []string           `json:"group_by_dimensions,omitempty"`
+	Filters           *MeterQueryFilters `json:"filters,omitempty"`
+}
+
+// MeterQueryRow is one aggregated bucket of a meter query result.
+type MeterQueryRow struct {
+	Value Numeric   `json:"value"`
+	From  time.Time `json:"from"`
+	To    time.Time `json:"to"`
+	// Dimensions holds the group-by values for this row. subject and customer_id
+	// are reserved dimension keys.
+	Dimensions map[string]string `json:"dimensions"`
+}
+
+// MeterQueryResult is the JSON result of a meter query.
+type MeterQueryResult struct {
+	From *time.Time      `json:"from,omitempty"`
+	To   *time.Time      `json:"to,omitempty"`
+	Data []MeterQueryRow `json:"data"`
+}
+
+// CreateMeterRequest is the body for creating a meter.
+type CreateMeterRequest struct {
+	Name        string            `json:"name"`
+	Key         string            `json:"key"`
+	Aggregation MeterAggregation  `json:"aggregation"`
+	EventType   string            `json:"event_type"`
+	Description *string           `json:"description,omitempty"`
+	Labels      map[string]string `json:"labels,omitempty"`
+	// EventsFrom, when set, is the date from which the meter includes events.
+	EventsFrom *time.Time `json:"events_from,omitempty"`
+	// ValueProperty is a JSONPath expression extracting the aggregated value from
+	// the event data. Ignored for count aggregation.
+	ValueProperty *string `json:"value_property,omitempty"`
+	// Dimensions maps group-by dimension names to JSONPath expressions.
+	Dimensions map[string]string `json:"dimensions,omitempty"`
+}
+
+// UpdateMeterRequest is the body for updating a meter. All fields are optional;
+// only the set ones are sent.
+type UpdateMeterRequest struct {
+	Name        *string           `json:"name,omitempty"`
+	Description *string           `json:"description,omitempty"`
+	Labels      map[string]string `json:"labels,omitempty"`
+	Dimensions  map[string]string `json:"dimensions,omitempty"`
+}
+
 // MetersService groups the meter operations. Access it via Client.Meters.
 type MetersService struct {
 	client *Client
-}
-
-// StringFilter expresses a comparison on a string field. Set exactly one
-// operator; unset fields are omitted from the request. It mirrors the API's
-// FilterString type. (The spec also allows a bare string shorthand for equality
-// on query-string filters, e.g. filter[key]=value; that is a server-side parse
-// convenience, not a distinct JSON shape — use Eq for the same effect.)
-type StringFilter struct {
-	Eq        *string  `json:"eq,omitempty"`
-	Neq       *string  `json:"neq,omitempty"`
-	Gt        *string  `json:"gt,omitempty"`
-	Gte       *string  `json:"gte,omitempty"`
-	Lt        *string  `json:"lt,omitempty"`
-	Lte       *string  `json:"lte,omitempty"`
-	Contains  *string  `json:"contains,omitempty"`
-	Oeq       []string `json:"oeq,omitempty"`
-	Ocontains []string `json:"ocontains,omitempty"`
-	Exists    *bool    `json:"$exists,omitempty"`
 }
 
 // MeterFilter narrows a meter listing by key and/or name.
 type MeterFilter struct {
 	Key  *StringFilter
 	Name *StringFilter
-}
-
-// PageParams selects a page of a paginated listing.
-type PageParams struct {
-	// Size is the number of items per page.
-	Size *int
-	// Number is the 1-based page number.
-	Number *int
 }
 
 // MeterListParams are the optional query parameters for listing meters. The
