+++ sdk/go/openmeter/types.go
@@ -1,0 +1,6 @@
+package openmeter
+
+// Numeric represents an arbitrary-precision number. The API encodes it as a
+// decimal string (e.g. "12.3456") to avoid float precision loss, so the SDK
+// surfaces it as a string. Parse with a decimal library when arithmetic is needed.
+type Numeric = string
